requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-retry==2.9.1
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
backports.zoneinfo==0.2.1
boto3==1.36.12
botocore==1.36.12
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2.18
django-redis==5.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
idna==3.10
jmespath==1.0.1
multidict==6.1.0
pillow==10.4.0
propcache==0.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
twilio==9.4.4
typing-extensions==4.12.2
urllib3==1.26.20
yarl==1.15.2
